Update/python base (#13)
* updating python base and pinning django version

pip, when upgraded, will install a version that is no longer supporting
python 3.5 because of a specific kind of format string. So here we want to upgrade
the python in the container to 3.8. I am also pinning the Django version because
the latest needs a bit of work to upgrade (and this can be done when needed in another PR)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* black formatting

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

Co-authored-by: vsoch <vsoch@users.noreply.github.com>